fix(connect): http errors handling for suite-native
